final submission

+++ templates/shaders/shaderFrag14.glsl
@@ -1,10 +1,12 @@
-precision lowp float;
+precision highp float;
 
 varying vec3 vColor;
 varying float vTotal;
 varying float vcurrentR;
 
 void main() {
+
+  // controling the number of layers with the trancparency
   if (vTotal >= vcurrentR) {
     gl_FragColor = vec4(vColor, 1.);
   } else {
